Refactor Numerical Answer Question Type
